mapper

Computes the mapper graph. Mapper is a tool for summarizing topological information from datasets and maps defined on them. It takes as input a set of 'point cloud' data, a (possibly lower dimensional) map defined on the data, and produces a topological summary of the data expressed through a cover equipped to the codomain of the map.

mapper(X, filter_values, cover_params = c(typename = "fixed rectangular",
  number_intervals = 10L, percent_overlap = 35), measure = "euclidean",
  cluster_params = c(cl = "single", num_bins = 10L),
  return_reference = FALSE)

Arguments

X
Either an \(n x D\) data matrix.

filter_values
An \(n x d\) data matrix.

cover_params
Named list of cover parameters. See details.

measure
String indicating the measure in the data space. Accepts any in pr_DB.

cluster_params
Named list of clustering parameters. See details.

return_reference
Boolean whether or not to return the reference class used to construct Mapper. See MapperRef.

Value

By default, returns a mapper object, which is a named list with members adjacency (adjacency matrix for the edges), vertices (list with point membership indices), and list level_sets (list with vertex indices). Each vertex in vertices also records the flat index of the level set that vertex intersects, recorded as an attribute.

If return_reference is TRUE, the MapperRef object is returned. This object can be used to fine-tune parts of the mapper construction.

Details

mapper is a generic function that concisely parameterizes the Mapper framework into a single function definition. This function serves as a convenience wrapper around the MapperRef R6 generator for users used to functional programming. For finer control over the mapper construction, use MapperRef instead. If return_reference is TRUE, the MapperRef instance used by this function is returned.

The cover_params must be a named list of all of the parameters needed by the cover. It must have a typename member matching one of the typenames listed in covers_available, along with all of the corresponding parameters for that cover type.

The cluster_params must be a named list containing members cl (string) and num_bins (integer), where cl is one of the linkage criterion used by hclust, and num_bins corresponds the binning argument in the cutting heuristic cutoff_first_bin.

References

Singh, Gurjeet, Facundo Memoli, and Gunnar E. Carlsson. "Topological methods for the analysis of high dimensional data sets and 3d object recognition." SPBG. 2007.

See also

Examples

data("noisy_circle", package="Mapper")
left_pt <- noisy_circle[which.min(noisy_circle[, 1]),]
f_x <- matrix(apply(noisy_circle, 1, function(pt) (pt - left_pt)[1]))

m <- Mapper::mapper(X = noisy_circle, filter_values = f_x,
                    cover_params = list(typename="restrained rectangular", number_intervals=10L, percent_overlap=50),
                    measure = "euclidean",
                    cluster_params = list(cl="single", num_bins=10L))