Constructs the k-skeleton

Computes the k-skeleton of the mapper by computing the nerve of the pullback of cover member.

Arguments

k
the maximal dimension to consider.

Details

The primary output of the Mapper method is a simplicial complex. With MapperRef objects, the simplicial complex is stored as a simplextree. The underlying complex does not need to be modified by the user, i.e. is completely maintained by MapperRef methods (e.g. this method, construct_nerve, etc.).

This function computes the k-skeleton inductively, e.g. by first computing the vertices, then the edges, etc. up to the dimension specified. A check is performed to ensure the pullback has been decomposed, and if not, then construct_pullback is called.

For an algorithmic description of this process, see Singh et. al, section 3.2.

References

Gurjeet Singh, Facundo Mémoli, and Gunnar Carlsson. "Topological methods for the analysis of high dimensional data sets and 3d object recognition." SPBG. 2007.