mapper

Computes the topological graph of the Mapper construction. Mapper is a tool for summarizing topological information from datasets and maps defined on them. It takes as input a set of 'point cloud' data and a (possibly low dimensional) coordinatization of the data, and produces a topological summary of the data expressed through a cover of the codomain of the map.

mapper(X, filter_values, cover_type = "fixed rectangular",
  return_reference = FALSE, ...)

Arguments

X
Either an \(n x m\) data matrix (preferred) or a dist object.

filter_values
An \(n x d\) matrix (or vector) of values produced .

cover_type
Name of the covering method to use.

return_reference
Boolean whether or not to return the reference class used to construct Mapper. See the return value section.

...
Additional arguments passed to other components in the Mapper framework. See details.

Details

mapper is a generic function, whose parameterization allows complete control on the resulting construction. Additional named parameters needed by any of these components may be passed through the three-dot ellipsis .... See the examples section for usage cases.

Additional arguments

number_intervals

Either a positive integer or a vector of m positive integers indicating the number of intervals to distribute along the filter space.

overlap

Either a single real-valued number or m-length vector of values between 0 and 1 specifying how much adjacent intervals should overlap.

measure

If X is a point cloud, string-value indicating the measure to compute on X. Ignored if X is a dist object.

cl

String-value of the linkage criterion to using for the hierarchical clustering step, or a clustering function which returns an integer vector indicating cluster membership. See details.

...

All named parameters are dispatched to other functional components of Mapper. See examples.

X may be a dist object or point cloud data in a matrix-coercible format (recommended).

The cl argument controls the clustering algorithm to use. If a string is given, any named linkage criterion in hclust may be used. Alternatively, a function may be given which accepts as input a dist object and returns an integer vector partitioning of the data. By default, if a hierarchical algorithm is used, a histogram-based heuristic is used to find the cutting threshold. The num_bins parameter may be given to use an alternative binning for the histogram (default is 10). See ?cutoff_first_bin for details.

The cover_type can be either "fixed rectangular" (default) or "restrained rectangular." Both covers rely on the same parameters, number_intervals and overlap.

If X is a point cloud and hierarchical clustering is used, distances must be computed on subsets of X. The measure parameter controls which distance measure to use. Supports any distance measure in parallelDist package. Alternatively, measure may be given as a function which accepts as input the subset of X to compute distances on.

All named parameters in ... are dispatched to their appropriate methods, and thus may be used to configure custom functions as well. For example, if a custom clustering function is supplied to the cl argument, named parameters in ... will be passed to the clustering algorithm as needed.

Return Value

By default, returns a mapper object, which is a named list with members adjacency (adjacency matrix for the edges), nodes (list with point membership indices), and list level_sets (list with node indices). Each node in nodes also records the flat index of the level set that node intersects, recorded as an attribute. The level_sets are ordered by their flat index, and named according to their multi-index.

If return_reference is TRUE, the MapperRef object is returned. This object can be used to fine-tune parts of the mapper construction.

References

Singh, Gurjeet, Facundo Memoli, and Gunnar E. Carlsson. "Topological methods for the analysis of high dimensional data sets and 3d object recognition." SPBG. 2007.

See also

Examples

# NOT RUN {
#install.packages("igraph")
library(igraph)
g1 <- graph.adjacency(m1$adjacency, mode="undirected")
plot(g1, layout = layout.auto(g1) )
# }