Assigns a distance measure

Assigns a distance measure to the MapperRef instance to use when compute distances in the data space.

Arguments

measure
The distance measure to use (string).

...
Extra parameters passed to the distance function. See details.

Details

The supplied measure must be either one of ["euclidean", "maximum", "manhattan", "canberra", "binary", "minkowski"], or, if the proxy package is installed, any measure in proxy::pr_DB$get_entry_names().

Additional parameters passed via ... are passed to either dist or, if installed, parallelDist. The former does not work with any measure in the proxy package, whereas the latter should.

Usage

$use_distance_measure(measure)

See also

Examples

data(noisy_circle)
m <- MapperRef$new(noisy_circle)
m$use_filter(noisy_circle[,1])
m$use_cover("fixed interval", number_intervals = 5, percent_overlap = 25)

## Constructs clusters with euclidean metric (default)
m$use_distance_measure("euclidean")
m$construct_pullback()

## Constructs clusters with p-norm (p = 1)
m$use_distance_measure("Minkowski", p = 1L)
m$construct_pullback()
# NOT RUN {
## To see list of available measures, use:
proxy::pr_DB$get_entry_names()
# }