Cover abstract class

Reference Class (R6) implementation of a Cover. This class is meant to act as an abstract class to derive other types of covering generators with. Minimally, a derived covering class must implement the 'construct_cover' method to populate the 'level_sets' list with point indices, and any parameters that the derived class requires.

Additional methods may also be added to improve the efficiency of the cover. See the vignette on creating a custom cover for details.

CoverRef

Format

An R6Class generator object

Fields

filter_values

(n x d) matrix of filter values

typename

Unique string identifier for the covering.

index_set

character vector used to index the 'level_sets' list

level_sets

list of the

Private variables

The following is a list of private variables available for derived classes. Each may be accessed by the private$ access method, see ?R6 for more details.

.level_setsnames list of indices in the original data set to cluster over.

.index_setcharacter vector of keys that uniquely index the level sets.

.filter_dimconstant representing the filter dimension.

.filter_sizeconstant representing the number of points in the filter space.

.typenamestring identifier of the covering method.